requirements.txt
PyAutoGUI~=0.9.54
huggingface-hub~=0.34.3
pytesseract~=0.3.13
pyperclip~=1.9.0
pynput~=1.8.1
langchain-core~=0.3.74
keyboard~=0.13.5
pillow